feat(bandit): repoint workflow pick to features /workflow-projection
features-service deleted GET /features/:slug/candidates and folded its
(audience x workflow) evidence into the reshaped GET
/features/:slug/workflow-projection (rows[] grain-ladder + resolved).

- rename src/lib/features-candidates-client.ts ->
  features-workflow-projection-client.ts; Candidate -> ProjectionRow,
  fetchCandidates -> fetchWorkflowProjectionRows
- fetch /workflow-projection, parse body.rows; send brandId + goal only
  (drop brandProfileId — the new endpoint derives economics from brandId)
- selectWorkflowGreedy ranks on row.resolved.costPerOutcomeUsd (argmin,
  null/<=0 skipped); audienceIdsForWorkflow keys on row.audienceId — both
  behaviors unchanged
- audience leg (/audience-stats) untouched
- update bandit/scheduler/internal-routes fixtures to the new row shape
