+++ CLAUDE.md
@@ -81,9 +81,11 @@ The campaign picks, per run, WHICH audience to contact and WHICH workflow to run
 **Cardinality (verified prod): 1 `/execute` = 1 start-run = 1 `lead-serve`.** The DAG does NOT loop start-run; each run is a brand-new `/execute` (scheduler `/end-run` → `nextRunAt` → next tick re-executes). So "vary the workflow per run" = pass a different slug to the NEXT execute — never a mid-flight DAG switch.
 
 **The two levers live at DIFFERENT code points, and that split is load-bearing:**
-- **Workflow** is chosen at the TRIGGER (`scheduler.ts reRunDueCampaigns`, before `executeCampaignWorkflow`) because the workflow is the DAG identity in the `/execute` URL — start-run is too late (the DAG is already running). `resolveWorkflowSlugForTrigger` greedy-picks (`selectWorkflowGreedy`) over features-service `/candidates` (per-workflow evidence). **Fail-soft**: falls back to `campaign.workflowSlug` when there's no evidence or features-service is down — a selection optimization must never block a run. Only `reRunDueCampaigns` triggers (not `claimStuckCampaigns`); campaign create/activate keep the configured slug for the seed run, the scheduler varies it from run 2.
-- **Audience** is chosen at `/start-run` (`selectAudienceForRun` Thompson over `/audience-stats`, all `active`, dropped `limit=1`) because workflow-service threads `x-audience-id` to downstream nodes from **start-run's RESPONSE** (`results.startRun.audienceId` in `dag-to-openflow.ts`), NOT from the execute input. start-run MUST keep returning `audienceId`. **Workflow-conditioning**: start-run fetches `/candidates` for the run's chosen workflow (`req.workflowSlug`), derives the audience-grain audienceIds via `audienceIdsForWorkflow` (rows with `audienceId != null` for that `workflowDynastySlug`), and passes them as `eligibleAudienceIds` to restrict the Thompson set. **Fail-soft**: any error, or no audience couples for that workflow (cold workflow) → no restriction, Thompson over all active audiences (a fresh workflow must still get an audience). Evidence + the full audience object (filters) still come from `/audience-stats`; `/candidates` only supplies the workflow→audience membership.
+- **Workflow** is chosen at the TRIGGER (`scheduler.ts reRunDueCampaigns`, before `executeCampaignWorkflow`) because the workflow is the DAG identity in the `/execute` URL — start-run is too late (the DAG is already running). `resolveWorkflowSlugForTrigger` greedy-picks (`selectWorkflowGreedy`) over features-service `/workflow-projection` `rows[]` (per-workflow evidence; ranks on `row.resolved.costPerOutcomeUsd`). **Fail-soft**: falls back to `campaign.workflowSlug` when there's no evidence or features-service is down — a selection optimization must never block a run. Only `reRunDueCampaigns` triggers (not `claimStuckCampaigns`); campaign create/activate keep the configured slug for the seed run, the scheduler varies it from run 2.
+- **Audience** is chosen at `/start-run` (`selectAudienceForRun` Thompson over `/audience-stats`, all `active`, dropped `limit=1`) because workflow-service threads `x-audience-id` to downstream nodes from **start-run's RESPONSE** (`results.startRun.audienceId` in `dag-to-openflow.ts`), NOT from the execute input. start-run MUST keep returning `audienceId`. **Workflow-conditioning**: start-run fetches `/workflow-projection` `rows[]` for the run's chosen workflow (`req.workflowSlug`), derives the audience-grain audienceIds via `audienceIdsForWorkflow` (rows with `audienceId != null` for that `workflowDynastySlug`), and passes them as `eligibleAudienceIds` to restrict the Thompson set. **Fail-soft**: any error, or no audience couples for that workflow (cold workflow) → no restriction, Thompson over all active audiences (a fresh workflow must still get an audience). Evidence + the full audience object (filters) still come from `/audience-stats`; `/workflow-projection` only supplies the workflow→audience membership.
 
-**Downstream honors the choice — already correct, don't rebuild:** lead-service `buffer/next` uses the passed `x-audience-id` and does NOT re-select; human-service suppression is **per-brand atomic, cross-audience** (the serve cursor is per-audience) so overlapping audiences never double-contact. `workflowSlug == workflowDynastySlug` (v1) so `/candidates`' dynasty slug is executable by `/by-slug/{slug}/execute`.
+**Downstream honors the choice — already correct, don't rebuild:** lead-service `buffer/next` uses the passed `x-audience-id` and does NOT re-select; human-service suppression is **per-brand atomic, cross-audience** (the serve cursor is per-audience) so overlapping audiences never double-contact. `workflowSlug == workflowDynastySlug` (v1) so the projection row's dynasty slug is executable by `/by-slug/{slug}/execute`.
 
-**Audience grain SHIPPED (features-service#368, prod v0.60.0; grain label renamed `persona`→`audience` in #376, prod v0.62.1):** `/candidates` now emits `grain:"audience"` rows with non-null `audienceId` per `(audienceId × workflowDynastySlug)` couple that has audience-attributed runs/outcomes. campaign-service keys on `audienceId != null` (NOT the grain string), so the rename is transparent to it. **Caveat — per-workflow OUTCOME splitting does NOT exist yet** (send/engagement untagged by workflow): an audience-grain row's outcome numbers are the audience-level slice (byte-identical to `/audience-stats`), the same across all of that audience's workflow couples. So `/candidates` reliably tells us WHICH audiences ran a workflow (the membership filter), but not how an audience performed *differently* per workflow. Full joint `P(a|w)` evidence (workflow-discriminated audience outcomes) still needs per-workflow outcome tagging upstream (features-service#366/#367, workflow-service#321). (Set 2026-06-23, campaign-service workflow-greedy + audience-conditioning; supersedes the 2026-06-21 #219 "not yet shipped" note.)
+**Audience grain SHIPPED (features-service#368, prod v0.60.0; grain label renamed `persona`→`audience` in #376, prod v0.62.1):** the projection `rows[]` carry non-null `audienceId` per `(audienceId × workflowDynastySlug)` couple that has audience-attributed runs/outcomes. campaign-service keys on `audienceId != null`. **Caveat — per-workflow OUTCOME splitting does NOT exist yet** (send/engagement untagged by workflow): an audience-grain row's outcome numbers are the audience-level slice (byte-identical to `/audience-stats`), the same across all of that audience's workflow couples. So the projection reliably tells us WHICH audiences ran a workflow (the membership filter), but not how an audience performed *differently* per workflow. Full joint `P(a|w)` evidence (workflow-discriminated audience outcomes) still needs per-workflow outcome tagging upstream (features-service#366/#367, workflow-service#321). (Set 2026-06-23, campaign-service workflow-greedy + audience-conditioning; supersedes the 2026-06-21 #219 "not yet shipped" note.)
+
+**Endpoint migration (2026-07-06):** features-service DELETED `/features/:slug/candidates` and folded its evidence into the reshaped `GET /features/:slug/workflow-projection` (`rows[]` grain-ladder + `resolved`). campaign-service reads `rows[]`, ranks on `row.resolved.costPerOutcomeUsd`, scopes audiences via `row.audienceId`. Client: `src/lib/features-workflow-projection-client.ts` (was `features-candidates-client.ts`). Sends `brandId` + `goal` only (brandProfileId dropped — the new endpoint derives economics from brandId).
