+++ src/routes/internal.ts
@@ -11,7 +11,7 @@ import { wakeScheduler } from "../lib/scheduler.js";
 import { traceEvent } from "../lib/trace-event.js";
 import { fetchBrandRuntimeContext } from "../lib/brand-runtime-client.js";
 import { selectAudienceForRun } from "../lib/features-audience-client.js";
-import { fetchCandidates, audienceIdsForWorkflow } from "../lib/features-candidates-client.js";
+import { fetchWorkflowProjectionRows, audienceIdsForWorkflow } from "../lib/features-workflow-projection-client.js";
 import type { DownstreamIdentity } from "../lib/downstream-headers.js";
 
 const router = Router();
@@ -210,23 +210,22 @@ router.post("/start-run", requireApiKey, requirePipelineHeaders, trackingHeaders
     // Scope the audience exploration to the audiences that have run under THIS run's
     // workflow (chosen greedily at the trigger), so we explore "the best audiences for
     // this workflow", not across every audience the brand ever contacted. Derived from
-    // features-service /candidates audience-grain rows. Fail-soft: any error or no audience-
-    // attributed couples for this workflow → no restriction (explore all active
+    // features-service /workflow-projection audience-grain rows. Fail-soft: any error or no
+    // audience-attributed couples for this workflow → no restriction (explore all active
     // audiences), so audience selection never blocks a run.
     let eligibleAudienceIds: string[] | undefined;
     try {
-      const candidates = await fetchCandidates({
+      const rows = await fetchWorkflowProjectionRows({
         featureSlug: featureSlug!,
         brandId: primaryBrandId,
         goal: brandRuntimeContext.currentGoal,
-        brandProfileId: brandRuntimeContext.brandProfile?.id,
         identity: preRunIdentity,
       });
-      const ids = audienceIdsForWorkflow(candidates, workflowSlug);
+      const ids = audienceIdsForWorkflow(rows, workflowSlug);
       if (ids.length > 0) eligibleAudienceIds = ids;
     } catch (err) {
       console.warn(
-        `[campaign-service] workflow-scoped audience candidates failed for brand ${primaryBrandId}, exploring all audiences:`,
+        `[campaign-service] workflow-scoped audience selection failed for brand ${primaryBrandId}, exploring all audiences:`,
         err,
       );
     }
